+++ .gitignore
@@ -1,1 +1,4 @@
 .DS_Store
+.byebug_history
+doc/
+.yardoc/
