+++ Gemfile
@@ -1,7 +1,4 @@
 source 'https://rubygems.org'
 
-gem "rake"
-gem "bundler"
-gem "json"
-gem "nokogiri"
-gem 'test-unit', :group => :development
+gemspec
+
