+++ lib/yahoo-finance/version.rb
@@ -1,4 +1,4 @@
 # Version Variable for Yahoo Finance Gem
 module YahooFinance
-  VERSION = '2.0.0'
+  VERSION = '3.0.0-alpha.1'
 end
