+++ test/test_helper.rb
@@ -1,0 +1,5 @@
+$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
+require 'yahoo-finance'
+
+require 'minitest/autorun'
+
